alert and notfication for new messages

+++ src/Chat.php
@@ -8,14 +8,16 @@ use Ratchet\ConnectionInterface;
 
 class Chat implements MessageComponentInterface {
 
-    private $clients;
-    protected $users;
+    protected $clients;
+    protected $usersImpl;
+    private $users;
     
-    public function __construct($users)
+    public function __construct($usersImpl)
     {
         // initialize clients storage
         $this->clients = new \SplObjectStorage;
-        $this->users = $users;
+        $this->usersImpl = $usersImpl;
+        $this->users = [];
     }
 
     public function onOpen(ConnectionInterface $conn)
@@ -23,39 +25,90 @@ class Chat implements MessageComponentInterface {
         // store new connection in clients
         $this->clients->attach($conn);
 
-        $voyants = $this->users->getAllUsers(['type'=>'V']);
-        // send a welcome message to the client that just connected
-        foreach ($this->clients as $client) {
-            $client->send(json_encode(array('type' => 'connect', 'text' => 'Welcome to the test chat app!', 'voyants' => $voyants)));
-        }
 
     }
 
+
+
     public function onClose(ConnectionInterface $conn)
     {
-        // remove connection from clients
-        $this->clients->detach($conn);
-        printf("Connection closed: %s\n", $conn->resourceId);
-        $voyants = $this->users->getAllUsers(['type'=>'V']);
-        foreach ($this->clients as $client) {
-            $client->send(json_encode(array('type' => 'disconnect','voyants' => $voyants)));
+        if(isset($this->users[$conn->resourceId])){
+            $user  = $this->users[$conn->resourceId] ;
+
+
+            // check if the user is alerday connected
+            $userConnected = array_filter(
+                $this->users,
+                function ($e) use ($user) {
+                    return $e->id == $user->id;
+                }
+            );
+
+            if(count($userConnected)<=1){
+                $this->updateUserStatus($user,0);
+            }
+            unset($this->users[$conn->resourceId]);
+            var_dump($user->username);
+            // remove connection from clients
+            $this->clients->detach($conn);
+
+            $voyants = $this->usersImpl->getAllUsers(['type'=>'V']);
+
+            foreach ($this->clients as $client) {
+                $client->send(json_encode(array('type' => 'disconnect','voyants' => $voyants)));
+            }
         }
+
     }
 
 
     public function onMessage(ConnectionInterface $conn, $message)
     {
         // send message out to all connected clients
-        foreach ($this->clients as $client) {
-            $client->send($message);
+        $data = json_decode($message);
+        // printf("Connection opened: %s\n");
+
+        switch($data->type){
+            case 'subscribe':
+
+                $this->usersImpl->setId($data->userId);
+                $user = $this->usersImpl->getUser();
+                if($user){
+                    $this->updateUserStatus($user,1);
+                    $this->users[$conn->resourceId] = $user ;
+                    // send a welcome message to the client that just connected
+                    $voyants = $this->usersImpl->getAllUsers(['type'=>'V']);
+                    $messages = $this->usersImpl->getAllMsg();
+
+                    foreach ($this->clients as $client) {
+                        $client->send(json_encode(array('type' => 'subscribe', 'voyants' => $voyants,'messages'=>$messages)));
+                    }
+                }
+                break;
+
+            default:
+
+                foreach ($this->clients as $client) {
+                    $client->send($message);
+                }
         }
 
     }
 
      public function onError(ConnectionInterface $conn, \Exception $e) {
         echo "An error has occurred: {$e->getMessage()}\n";
-
         $conn->close();
     }
+
+
+    public function updateUserStatus($user,int  $status){
+
+            $this->usersImpl->setId($user->id);
+            $this->usersImpl->setLoginStatus($status);
+           $this->usersImpl->updateLoginStatus();
+
+    }
+
+
 
 }
